externalizing TestSummary in valid.test_summary
